+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ## [1.9.1]
 
-### tokenization
+### motif
+#### tokenization
 ##### Added ✨
 
 ##### Changed 🔄
 
 ##### Fixed 🐛
-- Fixed `glycan_to_composition` handling of narrow monosaccharide modification wildcards such as `HexNAc4/6S`
+- Fixed `glycan_to_composition` handling of narrow monosaccharide modification wildcards such as `HexNAc4/6S` (b3ad039)
 
 ##### Deprecated ⚠️
+
+#### analysis
+##### Added ✨
+- `get_biodiversity` now has a new `circadian` API that supports analyzing whether changes in alpha-diversity (internally using `get_jtk`) and beta-diversity, using a distance-based redundancy analysis (db-RDA) regressing the Gower-centered dissimilarity matrix onto a cosinor design, are rhythmic